+++ app/src/scenes/Persons/FoldersNavigator.js
@@ -9,7 +9,8 @@ import colors from '../../utils/colors';
 import Documents from './Documents';
 import MedicalFile from './MedicalFile';
 import { useRecoilValue } from 'recoil';
-import { userState } from '../../recoil/auth';
+import { organisationState, userState } from '../../recoil/auth';
+import Group from './Group';
 
 const FoldersStack = createStackNavigator();
 
@@ -21,6 +22,7 @@ const FoldersNavigator = (props) => {
       <FoldersStack.Screen name="InformationsSocial">{(stackProps) => <InformationsSocial {...props} {...stackProps} />}</FoldersStack.Screen>
       <FoldersStack.Screen name="InformationsMedical">{(stackProps) => <InformationsMedical {...props} {...stackProps} />}</FoldersStack.Screen>
       <FoldersStack.Screen name="Documents">{(stackProps) => <Documents {...props} {...stackProps} />}</FoldersStack.Screen>
+      <FoldersStack.Screen name="Group">{(stackProps) => <Group {...props} {...stackProps} />}</FoldersStack.Screen>
       {!!user?.healthcareProfessional && (
         <FoldersStack.Screen name="MedicalFile">{(stackProps) => <MedicalFile {...props} {...stackProps} />}</FoldersStack.Screen>
       )}
@@ -32,12 +34,14 @@ export default FoldersNavigator;
 
 const FoldersSummary = ({ navigation, backgroundColor }) => {
   const user = useRecoilValue(userState);
+  const organisation = useRecoilValue(organisationState);
   return (
     <ScrollContainer noPadding backgroundColor={backgroundColor || colors.app.color}>
       <Spacer />
       <Row withNextButton caption="Informations sociales" onPress={() => navigation.navigate('InformationsSocial')} />
       <Row withNextButton caption="Informations médicales" onPress={() => navigation.navigate('InformationsMedical')} />
       <Row withNextButton caption="Documents" onPress={() => navigation.navigate('Documents')} />
+      {!!organisation.groupsEnabled && <Row withNextButton caption="Liens Familiaux" onPress={() => navigation.navigate('Group')} />}
       {!!user?.healthcareProfessional && (
         <>
           <Spacer />
